feat: add custom icons for Google and Github
The changes in this commit add custom icons for Google and Github. These icons are created using the Lucide React library. The Google icon is used to represent the Google login functionality, while the Github icon is used to represent the Github login functionality. These icons are exported from their respective files for use in other components.