feat: update metadata and homepage layout with new design and content
The metadata for the homepage has been updated to include a title and description. Additionally, the layout of the homepage has been enhanced with a new design featuring a navigation bar with logo, sign-in and sign-up buttons, and a section promoting the open-source infrastructure for digital learning. The design includes styling elements such as gradients, images, and buttons for GitHub star action.
feat: added favicon to admin and learner app