feat: signout api integration
Refactored the cookieDestroyer function to send a POST request to the LOGOUT_URL with the refresh token and authorization header. 
Upon successful logout, the function now deletes the "Authorization" and "Refresh" cookies and redirects to "/home".
feat: Add logout URL to authentication endpoints

Added a new constant for the logout URL in the authentication endpoints to 
handle user sign out functionality.
feat: Update redirect paths in cookieSetter and cookieDestroyer functions

Changed redirect paths in cookieSetter from "/admin/dashboard" to "/dashboard" and in cookieDestroyer from "/home" to "/". Added handling error message to prevent signout and token exposure.